+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Renamed `Form` type to `FormType`
+
 ## 0.4.0 (May 7, 2020)
 
 - Added `useFieldValue` hook